Create Code Assignment 2 Part 1
The code works without errors but my question to the instructor is how do I nest a function inside of another function using C#? The instructions for this part of the homework was to have a "sayHello" and a "goodbye" function within the main function but the above code is the only way I have been able to get this to work without causing some sort of parsing error. Going through the assigned reading I did find that a class could be nested within a class so is that how I would have to properly complete this portion of the homework? Thanks for your help.